+++ js/util.js
@@ -1,0 +1,3 @@
+export const getRandomNumber = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
+
+export const getRandomElement = (array) => array[Math.floor(Math.random() * array.length)];
